+++ src/main/java/pvpmode/api/common/version/EnumVersionComparison.java
@@ -1,0 +1,12 @@
+package pvpmode.api.common.version;
+
+/**
+ * An enum for the comparison of two versions.
+ * 
+ * @author CraftedMods
+ *
+ */
+public enum EnumVersionComparison
+{
+    CURRENT, OLDER, NEWER;
+}