+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed the level of the log indicating that a compatibility module wasn't loaded because of missing dependencies to debug
 * Unified the appearance of the chat GUIs
 * Tab-completions now only work for players which are permitted to use a certain command
+* Added a serverside, configurable version checker, which notifies operators about new versions
 
 #### Configuration:
 * Added `pvpconfig reload [configurationName]` which reloads a specific configuration or all configurations